+++ docs/fr_FR/Changelog.md
@@ -4,6 +4,7 @@
 -   Amélioration: Outil 'check_json.php' pour commandes inexistantes.
 -   Amélioration: Suppression cmds obsoletes 'presence.json' & 'presenceEP2.json'.
 -   Amélioration: Support Tuya PIR+24G_radar+light ZG-204ZM (2741).
+-   Amélioration: Outil 'check_json.php' pour subType 'slider'.
 
 ## 250115-BETA-2, ## 250115-BETA-1